Move copyToClipboard and removeLocalIds to utilities
