+++ src/utilities/index.ts
@@ -12,6 +12,7 @@ export {default as formatters} from './form-helpers/formatters';
 export {default as dispatchers} from './redux-helpers/dispatchers';
 export {default as validators} from './form-helpers/validators';
 export {default as hints} from './form-helpers/hints';
+export {default as copyToClipboard} from './form-helpers/copy-to-clipboard';
 
 export {default as fetchSurveys} from './ajax-helpers/fetching';
 export {default as generateValidOAuthToken} from './auth-helpers/jwt-handling';
@@ -21,5 +22,6 @@ export {surveyGetRequest} from './ajax-helpers/axios-clients';
 
 export {default as fieldTemplate} from './template-helpers/add-field';
 export {default as optionTemplate} from './template-helpers/add-option';
+export {default as removeLocalIds} from './template-helpers/remove-local-ids';
 
 export {default as colors} from './ui-helpers/colors';
